Collision
Added a collider to the fisherman, but now the fisherman starts spinning
when he hits an object.
I have made the score system work. Moved the collision-code to the
fisherman-script instead. Don't know if that is a bad idea (now the
other scripts are practically empty). I have also made some small
changes to the score-GUI.