Hero
I’ve now made a Start scene, build it with our Main scene, to be the
new ‘Gameplay’, where you can try out the game..

I’ve also made the camera as child to the Fisherman, so it follows only
with the x-axis and not with the jump, so it is possible to se if it’s
hitting other obstacles

At last I made the jump function! Yaaay!

I also deleted the Maincamera.cs-script, it will not be needed anymore